Move SLO definition parsing code to `dashboard` package
