+++ internal/sli/dashboard/key_value_parsing.go
@@ -1,0 +1,11 @@
+package dashboard
+
+import "fmt"
+
+type duplicateKeyError struct {
+	key string
+}
+
+func (err *duplicateKeyError) Error() string {
+	return fmt.Sprintf("duplicate key '%s' in SLO definition", err.key)
+}
